requirements.txt
absl-py==0.8.1
affine==2.3.0
alabaster==0.7.12
anaconda-client==1.2.2
asn1crypto==1.0.1
astor==0.8.0
astroid==2.3.1
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.3.4
boto==2.49.0
boto3==1.10.5
botocore==1.13.5
Bottleneck==1.2.1
bumpversion==0.5.3
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.2.2
clyent==1.2.1
colorama==0.4.1
conda==4.3.16
configparser==3.7.3
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.7
descartes==1.1.0
distributed==2.5.2
docutils==0.15.2
earthpy==0.8.0
entrypoints==0.3
ephem==3.7.7.0
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Fiona==1.8.11
Flask==1.1.1
Flask-Cors==3.0.8
fsspec==0.5.2
future==0.17.1
gast==0.2.2
geojson==2.5.0
geopandas==0.6.2
gevent==1.4.0
glob2==0.7
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.8
imageio==2.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
msgpack==0.6.1
munch==2.5.0
networkx==2.4
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.1
Pillow==6.2.0
protobuf==3.11.2
psutil==5.6.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pymongo==3.10.1
pyparsing==2.4.2
pyproj==2.4.0
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
rasterio==1.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Rtree==0.9.1
ruamel.yaml==0.15.46
s3transfer==0.2.1
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.4.1
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
snuggs==1.4.7
sortedcontainers==2.1.0
soupsieve==1.9.3
tblib==1.4.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.6
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zict==1.0.0
